Generate TS files
