--- maas-schemas-ts/src/core/components/reason.ts
@@ -1,45 +1,0 @@
-/*
-
-undefined
-MaaS error reason definition
-
-!!! AUTO GENERATED BY CONVERT.TS REFRAIN FROM MANUAL EDITING !!!
-
-*/
-
-import * as t from 'io-ts';
-
-export const schemaId = 'http://maasglobal.com/core/components/reason.json';
-
-// Reason
-// The default export. More information at the top.
-export type Reason = t.Branded<
-  {
-    text?: string;
-    errorCode?: string | number;
-  },
-  ReasonBrand
->;
-export const Reason = t.brand(
-  t.partial({
-    text: t.string,
-    errorCode: t.union([t.string, t.number]),
-  }),
-  (
-    x,
-  ): x is t.Branded<
-    {
-      text?: string;
-      errorCode?: string | number;
-    },
-    ReasonBrand
-  > => true,
-  'Reason',
-);
-export interface ReasonBrand {
-  readonly Reason: unique symbol;
-}
-
-export default Reason;
-
-// Success
