+++ maas-schemas-ts/translation.log
@@ -242,6 +242,12 @@ WARNING: minLength field not supported outside top-level definitions
   in ./core/components/customerSelection.json
 WARNING: maxLength field not supported outside top-level definitions
   in ./core/components/customerSelection.json
+INFO: primitive type "string" used outside top-level definitions
+  in ./core/components/errors.json
+INFO: primitive type "string" used outside top-level definitions
+  in ./core/components/errors.json
+INFO: primitive type "number" used outside top-level definitions
+  in ./core/components/errors.json
 WARNING: minimum field not supported outside top-level definitions
   in ./core/components/fare.json
 INFO: primitive type "string" used outside top-level definitions
@@ -412,12 +418,6 @@ WARNING: minLength field not supported outside top-level definitions
   in ./core/components/place.json
 WARNING: maxLength field not supported outside top-level definitions
   in ./core/components/place.json
-INFO: primitive type "string" used outside top-level definitions
-  in ./core/components/reason.json
-INFO: primitive type "string" used outside top-level definitions
-  in ./core/components/reason.json
-INFO: primitive type "number" used outside top-level definitions
-  in ./core/components/reason.json
 INFO: primitive type "string" used outside top-level definitions
   in ./core/components/state-log.json
 INFO: primitive type "string" used outside top-level definitions
